Use authentication in tests
