+++ test/mocha.opts
@@ -1,2 +1,3 @@
 --recursive
 --require babel-register
+--timeout 5000
